pyproject.toml
[tool.poetry]
name = "lab1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
python-multipart = "^0.0.19"
uvicorn = "^0.34.0"
httpie = "^3.2.4"
python-slugify = "^8.0.4"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
